refactor: features 레이어 컴포넌트 내부 함수 lib로 이동
