+++ src/refactoring/features/cart/lib/cartUtils.ts
@@ -1,0 +1,13 @@
+import { ICartItem } from '../../../shared/types';
+
+export const getAppliedDiscount = (item: ICartItem) => {
+  const { discounts } = item.product;
+  const { quantity } = item;
+  let appliedDiscount = 0;
+  for (const discount of discounts) {
+    if (quantity >= discount.quantity) {
+      appliedDiscount = Math.max(appliedDiscount, discount.rate);
+    }
+  }
+  return appliedDiscount;
+};
