+++ src/refactoring/features/cart/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './cartUtils.ts';
