+++ src/refactoring/features/cart/ui/CartItem.tsx
@@ -1,4 +1,5 @@
 import { ICartItem } from '../../../shared/types';
+import { getAppliedDiscount } from '../lib';
 
 interface CartItemProps {
   item: ICartItem;
@@ -11,18 +12,6 @@ export function CartItem({
   updateQuantity,
   removeFromCart,
 }: CartItemProps) {
-  const getAppliedDiscount = (item: ICartItem) => {
-    const { discounts } = item.product;
-    const { quantity } = item;
-    let appliedDiscount = 0;
-    for (const discount of discounts) {
-      if (quantity >= discount.quantity) {
-        appliedDiscount = Math.max(appliedDiscount, discount.rate);
-      }
-    }
-    return appliedDiscount;
-  };
-
   const appliedDiscount = getAppliedDiscount(item);
 
   return (
