+++ src/refactoring/features/stock/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './stockUtils.ts';
